+++ src/infrastructure/repository/index.ts
@@ -3,6 +3,7 @@ export { parseFinderColumns, parseFinderMethod } from './parser';
 export type {
   Repository,
   ExtendableRepository,
+  MappedRepository,
   BaseRepository,
   SimpleFinders,
   MultiFinders,
@@ -14,9 +15,13 @@ export type {
   QueryOptions,
   CreateManyOptions,
   CreateManyCountResult,
+  Transformer,
+  MappedInsertable,
   RawBuilder,
   SelectQueryBuilder,
   Selectable,
   Insertable,
   Updateable,
+  Unwrap,
+  UnwrapRow,
 } from './types';